Refactor KafkaMessageType over to model module
