+++ src/main/java/com/rackspace/salus/telemetry/ambassador/services/KafkaEgress.java
@@ -19,7 +19,7 @@
 package com.rackspace.salus.telemetry.ambassador.services;
 
 import com.rackspace.salus.telemetry.ambassador.config.AmbassadorProperties;
-import com.rackspace.salus.telemetry.ambassador.types.KafkaMessageType;
+import com.rackspace.salus.telemetry.messaging.KafkaMessageType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
